Cyber Alien lobbing sprites added
Discovered the missing attack animation for boss 3 hidden away in the res wad file. Added those in and updated the enemy script to use it.
Also updated Tile014 to remove redundant switch sprites, the cutscene script was adjusted to fix a warning, and a Gianni sound clip mistakenly not pushed before has been added.